breaking: Renamed profile upload/delete API
