+++ jasao-server/src/main/kotlin/io/yeahx4/jasao/controller/file/UploadedFileController.kt
@@ -27,79 +27,7 @@ import org.springframework.web.multipart.MultipartFile
 class UploadedFileController(
     private val jwtService: JwtService,
     private val uploadedFileService: UploadedFileService,
-    private val userService: UserService
+    private val userService: UserService,
 ) {
     private val logger = LoggerFactory.getLogger(this::class.java)
-
-    /**
-     * Upload file
-     *
-     * # Request
-     * ### HTTP
-     * POST `/file/auth/upload`
-     *
-     * ### param
-     * - role : "profile"
-     *
-     * ### body
-     * - MultipartFile
-     *
-     * # Response
-     * - 400 : Invalid role
-     * - 200 : Success
-     */
-    @PostMapping("/auth/upload")
-    @Transactional
-    fun uploadImage(
-        @RequestHeader("Authorization") jwt: String,
-        @RequestParam role: String,
-        @RequestBody file: MultipartFile
-    ): Res<String> {
-        val user = this.jwtService.getUserFromToken(jwt)
-
-        return if (role == "profile") {
-            val path = this.uploadedFileService.saveProfileImage(user.id, file)
-            val dbUser = userService.getUserById(user.id)!!
-            dbUser.profile = path
-
-            this.logger.info("Profile image upload by user ${user.id}")
-
-            Res(HttpResponse("Ok", path), HttpStatus.OK)
-        } else {
-            this.logger.warn("Invalid role(${role}) image upload request by user ${user.id}")
-
-            Res(HttpResponse("Invalid role", null), HttpStatus.BAD_REQUEST)
-        }
-    }
-
-    /**
-     * Delete uploaded file
-     *
-     * # Request
-     * ### HTTP
-     * DELETE `/file/auth/profile/delete`
-     *
-     * ### Header
-     * - Authorization: JWT
-     *
-     * # Response
-     * - 400 : Image not exists
-     * - 200 : Success
-     */
-    @DeleteMapping("/auth/profile/delete")
-    @Transactional
-    fun deleteProfileImage(@RequestHeader("Authorization") jwt: String): Res<String> {
-        val user = this.jwtService.getUserFromToken(jwt)
-        val dbUser = this.userService.getUserById(user.id)!!
-
-        if (!this.uploadedFileService.deleteProfileImage(user.id, user.profile.endsWith(".jpg"))) {
-            this.logger.warn("User ${user.id} tried to delete non-exist profile image")
-            return Res(HttpResponse("File not exists", null), HttpStatus.BAD_REQUEST)
-        } else {
-            dbUser.profile = ""
-        }
-
-        this.logger.info("User ${user.id} removed its profile image")
-        return Res(HttpResponse("Ok", null), HttpStatus.OK)
-    }
 }
